14th commit

+++ src/main/java/rock/service/AnswerService.java
@@ -5,6 +5,8 @@ import org.springframework.stereotype.Service;
 
 import rock.domain.Answer;
 import rock.domain.AnswerRepository;
+import rock.domain.Question;
+import rock.domain.User;
 
 @Service
 public class AnswerService {
@@ -13,7 +15,12 @@ public class AnswerService {
 	@Autowired
 	AnswerRepository answerRepository;
 
-	public void save(Answer ans) {
+	@Autowired
+	private QuestionService questionService;
+	
+	public void save(Answer ans, Long questionid, User user) {
+		Question qus = questionService.findOne(questionid);
+		ans.settingDBData(qus, user);
 		answerRepository.save(ans);
 		
 	}
@@ -23,8 +30,11 @@ public class AnswerService {
 		return answerRepository.findOne(id);
 	}
 
-	public void delete(Long id) {
-		answerRepository.delete(id);
+	public void delete(Long id, User user) {
+		Answer ans = findOne(id);
+		if(ans.userMatching(user)){
+			answerRepository.delete(id);
+		}
 		
 	}
